# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "employee": {
                      "id": "uuid",
                      "email": "admin@hotel.com",
                      "firstName": "Admin",
                      "lastName": "User",
                      "role": "ADMIN"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Employee login",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout successful"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/v1/auth/me": {
      "post": {
        "operationId": "AuthController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user info",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "uuid",
                    "email": "admin@hotel.com",
                    "firstName": "Admin",
                    "lastName": "User",
                    "role": "ADMIN"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user info",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/v1/customers": {
      "post": {
        "operationId": "CustomersController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "Email or document already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new customer",
        "tags": [
          "Customers"
        ]
      },
      "get": {
        "operationId": "CustomersController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentType",
            "required": false,
            "in": "query",
            "description": "Filter by document type",
            "schema": {
              "type": "string",
              "enum": [
                "PASSPORT",
                "NATIONAL_ID",
                "DRIVERS_LICENSE",
                "OTHER"
              ]
            }
          },
          {
            "name": "nationality",
            "required": false,
            "in": "query",
            "description": "Filter by nationality",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all customers with pagination and filters",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/v1/customers/{id}": {
      "get": {
        "operationId": "CustomersController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer details with recent reservations"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get customer by ID",
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update customer",
        "tags": [
          "Customers"
        ]
      },
      "delete": {
        "operationId": "CustomersController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer deleted successfully"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete customer (soft delete)",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/v1/customers/{id}/reservations": {
      "get": {
        "operationId": "CustomersController_getReservationHistory_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer reservation history"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get customer reservation history",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/v1/customers/{id}/restore": {
      "post": {
        "operationId": "CustomersController_restore_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer restored successfully"
          },
          "404": {
            "description": "Deleted customer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore deleted customer",
        "tags": [
          "Customers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hotel Management API",
    "description": "API for hotel management system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@hotel.com",
            "description": "Employee email"
          },
          "password": {
            "type": "string",
            "example": "Admin123!",
            "description": "Employee password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name"
          },
          "email": {
            "type": "string",
            "example": "john.doe@email.com",
            "description": "Email address"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890",
            "description": "Phone number"
          },
          "documentType": {
            "type": "string",
            "enum": [
              "PASSPORT",
              "NATIONAL_ID",
              "DRIVERS_LICENSE",
              "OTHER"
            ],
            "description": "Document type"
          },
          "documentNumber": {
            "type": "string",
            "example": "AB123456",
            "description": "Document number"
          },
          "address": {
            "type": "string",
            "example": "123 Main St",
            "description": "Address"
          },
          "nationality": {
            "type": "string",
            "example": "USA",
            "description": "Nationality"
          },
          "birthDate": {
            "type": "string",
            "example": "1990-01-15",
            "description": "Birth date"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "documentType",
          "documentNumber"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name"
          },
          "email": {
            "type": "string",
            "example": "john.doe@email.com",
            "description": "Email address"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890",
            "description": "Phone number"
          },
          "documentType": {
            "type": "string",
            "enum": [
              "PASSPORT",
              "NATIONAL_ID",
              "DRIVERS_LICENSE",
              "OTHER"
            ],
            "description": "Document type"
          },
          "documentNumber": {
            "type": "string",
            "example": "AB123456",
            "description": "Document number"
          },
          "address": {
            "type": "string",
            "example": "123 Main St",
            "description": "Address"
          },
          "nationality": {
            "type": "string",
            "example": "USA",
            "description": "Nationality"
          },
          "birthDate": {
            "type": "string",
            "example": "1990-01-15",
            "description": "Birth date"
          }
        }
      },
...
    }
  }
}